// app/models/report_districts.ts
export type ReportDistricts = {
  REPORT_DISTRICT_ID?: number;
  REPORT_DISTRICT_CODE: string | null;
  SERVICE_TYPE: string | null;
  COMMUNITY_TYPE: string | null;
  NAME: string | null;
  LABEL: string | null;
  AGENCY: string | null;
  STATION_NAME: string | null;
  STATION: string | null;
  OMEGA_LABEL: string | null;
  OMEGA_NAME: string | null;
  SHAPE_LENGTH: number | null;
  SHAPE_AREA: number | null;
};